Separeted SavedModelVariablesIndex class from Saved Model
